moved FilterByContent down
